+++ clare/clare/retry/__init__.py
@@ -1,7 +1,8 @@
 # -*- coding: utf-8 -*-
 
 from .attempt import Attempt
-from .stop_strategies import AfterAttempt
+from .stop_strategies import AfterAttempt, AfterDuration
 
 __all__ = ['AfterAttempt',
+           'AfterDuration',
            'Attempt']