+++ clare/documentation/release_notes.md
@@ -6,7 +6,8 @@
 - Fixed parsing null log messages (`ME-648`).
 
 ##### Features
-- Added stopping after *n* attempts (`ME-659`).
+- Added stopping after attempt (`ME-659`).
+- Added stopping after duration (`ME-659`).
 
 ## v0.2.0 (Pending)
 ##### Changes